requirements.txt
asgiref==3.5.2
certifi==2022.6.15
distlib==0.3.5
dj-database-url==1.0.0
Django==4.0.6
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-on-heroku==1.1.2
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
filelock==3.7.1
gunicorn==20.1.0
Pillow==9.2.0
pipenv==2022.7.4
platformdirs==2.5.2
psycopg2-binary==2.9.3
pymongo==3.12.0
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.2.4
tzdata==2022.1
virtualenv==20.15.1
virtualenv-clone==0.5.7
whitenoise==6.2.0
